+++ apps/web/src/app/[locale]/(space)/layout.tsx
@@ -7,8 +7,8 @@ import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
 import { LicenseLimitWarning } from "@/features/licensing/components/license-limit-warning";
 import { CommandMenu } from "@/features/navigation/command-menu";
 import { isUserOnboarded } from "@/features/setup/utils";
-import { TimezoneProvider } from "@/features/timezone/client/context";
 import { AbilityProvider } from "@/lib/ability-manager/client";
+import { TimezoneProvider } from "@/lib/timezone/client/context";
 import { SpaceSidebar } from "./components/sidebar/space-sidebar";
 import { SpaceSidebarProvider } from "./components/sidebar/space-sidebar-provider";
 import { TopBar, TopBarLeft, TopBarRight } from "./components/top-bar";
