+++ apps/web/src/features/scheduled-event/components/scheduled-event-list.tsx
@@ -14,8 +14,8 @@ import { MoreHorizontalIcon, XIcon } from "lucide-react";
 import { ParticipantAvatarBar } from "@/components/participant-avatar-bar";
 import { StackedList } from "@/components/stacked-list";
 import { Trans } from "@/components/trans";
-import { FormattedDateTime } from "@/features/timezone/client/formatted-date-time";
 import { useSafeAction } from "@/lib/safe-action/client";
+import { FormattedDateTime } from "@/lib/timezone/client/formatted-date-time";
 import { cancelEventAction } from "../actions";
 
 export const ScheduledEventList = StackedList;
