+++ apps/web/src/features/setup/components/setup-form.tsx
@@ -16,9 +16,9 @@ import { useForm } from "react-hook-form";
 import { LanguageSelect } from "@/components/poll/language-selector";
 import { TimeZoneSelect } from "@/components/time-zone-picker/time-zone-select";
 import { Trans } from "@/components/trans";
-import { useTimezone } from "@/features/timezone";
 import { useTranslation } from "@/i18n/client";
 import { useSafeAction } from "@/lib/safe-action/client";
+import { useTimezone } from "@/lib/timezone";
 
 import { completeSetupAction } from "../actions";
 import type { SetupFormValues } from "../schema";
